requirements.txt
azure-cognitiveservices-speech
beautifulsoup4
ipykernel
openai==1.30.5
polars
python-dotenv
requests
wheel